Refactored forward computation code from 'batch_norm_op.cu'

+++ python/paddle/fluid/tests/unittests/test_batch_norm_op.py
@@ -440,7 +440,7 @@ class TestBatchNormOpTraining(unittest.TestCase):
                     "SavedMean": block.var('saved_mean'),
                     "SavedVariance": block.var('saved_variance')
                 }
-                block.create_var(name="reserve_space", dtype='float32')
+                block.create_var(name="reserve_space", dtype='float16')
                 outputs["ReserveSpace"] = block.var('reserve_space')
                 # has_reserve_space = False
                 # if data_format == 'NHWC':
